+++ test/iterator-recursion-test.js
@@ -1,4 +1,4 @@
-const test          = require('tap').test
+const test          = require('tape')
     , testCommon    = require('abstract-leveldown/testCommon')
     , leveldown     = require('../')
     , child_process = require('child_process') 
@@ -32,20 +32,19 @@ test('try to create an iterator with a blown stack', function (t) {
   // Reducing the stack size down from the default 984 for the child node
   // process makes it easier to trigger the bug condition. But making it too low
   // causes the child process to die for other reasons.
-  var opts  = { execArgv: ["--stack-size=128"] }
-  ,   child = child_process.fork(__dirname + '/stack-blower.js', ["run"], opts)
+  var opts  = { execArgv: [ '--stack-size=128' ] }
+    , child = child_process.fork(__dirname + '/stack-blower.js', [ 'run' ], opts)
+
+  t.plan(2)
   
   child.on('message', function (m) {
-      t.ok(true, m)
-      child.disconnect()
-      
-      t.end()
-    })
-    .on('exit', function (code, sig) {
-      t.ok(false, "Child exited with code=" + code + " sig=" + sig)
+    t.ok(true, m)
+    child.disconnect()
+  })
 
-      t.end()
-    })
+  child.on('exit', function (code, sig) {
+    t.equal(code, 0, 'child exited normally')
+  })
 })
 
 test('iterate over a large iterator with a large watermark', function (t) {